+++ lldb/include/lldb/Host/posix/HostInfoPosix.h
@@ -35,6 +35,12 @@ public:
 
   static UserIDResolver &GetUserIDResolver();
 
+#ifdef LLDB_ENABLE_SWIFT
+  static FileSpec GetSwiftResourceDir();
+  static bool ComputeSwiftResourceDirectory(FileSpec &lldb_shlib_spec,
+                                            FileSpec &file_spec, bool verify);
+#endif
+  
 protected:
   static bool ComputeSupportExeDirectory(FileSpec &file_spec);
   static bool ComputeHeaderDirectory(FileSpec &file_spec);
