+++ lldb/source/Host/macosx/objcxx/CMakeLists.txt
@@ -1,3 +1,8 @@
+set(SWIFT_SOURCES HostInfoMacOSXSwift.cpp)
+set(LLVM_OPTIONAL_SOURCES ${SWIFT_SOURCES})
+if (NOT LLDB_ENABLE_SWIFT_SUPPORT)
+  unset(SWIFT_SOURCES)
+endif()
 
 remove_module_flags()
 include_directories(..)
@@ -6,6 +11,9 @@ add_lldb_library(lldbHostMacOSXObjCXX
   Host.mm
   HostInfoMacOSX.mm
   HostThreadMacOSX.mm
+  HostInfoMacOSXSwift.cpp
+
+  ${SWIFT_SOURCES}
 
   LINK_LIBS
     lldbUtility
