+++ lldb/source/Host/posix/HostInfoPosix.cpp
@@ -1,4 +1,4 @@
-//===-- HostInfoPosix.cpp -------------------------------------------------===//
+//===-- HostInfoPosixSwift.cpp --------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
