+++ lldb/source/Plugins/ExpressionParser/Swift/CMakeLists.txt
@@ -7,7 +7,6 @@ add_lldb_library(lldbPluginExpressionParserSwift PLUGIN
   SwiftExpressionParser.cpp
   SwiftExpressionSourceCode.cpp
   SwiftExpressionVariable.cpp
-  SwiftHost.cpp
   SwiftPersistentExpressionState.cpp
   SwiftREPL.cpp
   SwiftREPLMaterializer.cpp
