+++ lldb/unittests/Expression/CMakeLists.txt
@@ -1,9 +1,3 @@
-set(SWIFT_SOURCES SwiftParserTest.cpp)
-set(LLVM_OPTIONAL_SOURCES ${SWIFT_SOURCES})
-if (NOT LLDB_ENABLE_SWIFT_SUPPORT)
-  unset(SWIFT_SOURCES)
-endif()
-
 add_lldb_unittest(ExpressionTests
   ClangParserTest.cpp
   ClangExpressionDeclMapTest.cpp
@@ -11,8 +5,6 @@ add_lldb_unittest(ExpressionTests
   DWARFExpressionTest.cpp
   CppModuleConfigurationTest.cpp
   ClangParserTest.cpp
-
-  ${SWIFT_SOURCES}
 
   LINK_LIBS
     lldbCore
